requirements.txt
pandas
langchain-google-genai
langchain-community
google-generativeai
python-dotenv
langchain-experimental
tabulate
streamlit
PyPDF2
